Add Deploy::Configuration and Deploy::Debug.
Deploy::Debug provides a wrapper for your response, opts, config and command... allowing you to get deeper insight into how we built your request and sent it out.

Deploy::Configuration is wrapper that encapsulates the reading and storage/modification of config.  It uses Forwardable to forward to the hash in an efficient way without much trouble.